Adjust trainer.rs to the async assembly way

+++ neurs/src/train/interface.rs
@@ -32,5 +32,5 @@ where
         &mut self,
         assembly: &mut AssemblyType,
         assembly_frame: &mut ATF,
-    ) -> Result<f64, String>;
+    ) -> Result<f32, String>;
 }
